+++ R/utilities-enum.R
@@ -1,7 +1,8 @@
 #' @title enumeration keys for Time range shortcut
 #' @export
 #' @family enumerations
-TIMERANGE <- ospsuite.utils::enum(c( # nolint
+TIMERANGE <- ospsuite.utils::enum(c(
+  # nolint
   total = "total",
   firstApplication = "firstApplication",
   lastApplication = "lastApplication"
@@ -10,18 +11,21 @@ TIMERANGE <- ospsuite.utils::enum(c( # nolint
 #' @title enumeration keys for `DataClass`
 #' @export
 #' @family enumerations
-DATACLASS <- ospsuite.utils::enum(c( # nolint
+DATACLASS <- ospsuite.utils::enum(c(
+  # nolint
   tpIndividual = "tp Individual",
   tpAggregated = "tp Aggregated",
   pkIndividual = "pk Individual",
   pkAggregated = "pk Aggregated",
-  tpTwinPop = "tp Virtual Twin Population"
+  tpTwinPop = "tp Virtual Twin Population",
+  numericValues = "numeric Values"
 ))
 
 #' @title enumeration keys for exportDirectories
 #' @export
 #' @family enumerations
-EXPORTDIR <- ospsuite.utils::enum(c( # nolint
+EXPORTDIR <- ospsuite.utils::enum(c(
+  # nolint
   simulationResult = "SimulationResults",
   sensitivityResults = "SensitivityResults",
   pKAnalysisResults = "PKAnalysisResults"
@@ -31,7 +35,8 @@ EXPORTDIR <- ospsuite.utils::enum(c( # nolint
 #' @title enumeration biometric units
 #' @export
 #' @family enumerations
-BIOMETRICUNITS <- ospsuite.utils::enum(c( # nolint
+BIOMETRICUNITS <- ospsuite.utils::enum(c(
+  # nolint
   age = "year(s)",
   weight = "kg",
   height = "cm"